+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `AppDomainDisposable` internal class to create a new `AppDomain` and dispose it.
 - Add `TryChangeType` to try to convert `object` to `ParameterType`.
 - Update `GetTestNameWithArguments` to support generic types.
+- Update `ValueToArgumentName` to support `double` and `float`.
 ### Tests
 - Add `SampleMetadata` and `SampleMetadataAttribute` test project.
 - Add `TestParameterCaseDouble` to test parameter with `double` type and parameter `int`. (Fix: #29)